requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-bootstrap-v5==1.0.11
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-environ==0.11.2
django-jalali-date==1.0.2
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-modeltrans==0.7.5
django-phone-verify==0.1.0
django-phonenumber-field==7.1.0
django-user-language-middleware==0.0.3
filelock==3.12.3
idna==3.4
jdatetime==4.1.1
phonenumbers==8.13.19
Pillow==10.0.0
python-gettext==5.0
requests==2.31.0
requests-file==1.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tldextract==3.4.4
typing_extensions==4.7.1
urllib3==2.0.4
